+++ priorityqueue.h
@@ -5,12 +5,12 @@
 #include <QPoint>
 #include <algorithm>
 
-class PriorityQueue : public QVector<QPair<int, QPoint>>
+class PriorityQueue : public QVector<QPair<float, QPoint>>
 {
 public:
     PriorityQueue();
     void place_elem(); //помещение элемента на правильное место
-    void push_back(QPair<int, QPoint>); //переопределение добавления элемента в конец
+    void push_back(QPair<float, QPoint>); //переопределение добавления элемента в конец
 };
 
 #endif // PRIORITYQUEUE_H
